requirements.txt
discord-webhook
python-dotenv
iconsdk
